# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EDU SMART OS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [finance]
      summary: List invoices
      parameters:
        - name: month
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [finance]
      summary: Create a monthly invoice batch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Created invoice batch
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    patch:
      operationId: updateInvoice
      tags: [finance]
      summary: Update invoice status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceUpdate"
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /expenses:
    get:
      operationId: listExpenses
      tags: [finance]
      summary: List operational expenses
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [finance]
      summary: Create an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /finance/summary:
    get:
      operationId: getFinanceSummary
      tags: [finance]
      summary: Get financial summary (revenue, expenses, profit)
      parameters:
        - name: month
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Financial summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinanceSummary"
...
components:
  schemas:
...
    Invoice:
      type: object
      required: [id, student_id, month, amount, status]
      properties:
        id: { type: string }
        student_id: { type: string }
        student_name: { type: ["string", "null"] }
        month: { type: string }
        amount: { type: number }
        status: { type: string }
        is_exempt: { type: boolean }
        paid_at: { type: ["string", "null"] }
        created_at: { type: string }
    InvoiceInput:
      type: object
      required: [month]
      properties:
        month: { type: string }
        amount_override: { type: number }
    InvoiceUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string }
        paid_at: { type: string }
    Expense:
      type: object
      required: [id, category, amount, date]
      properties:
        id: { type: string }
        category: { type: string }
        description: { type: ["string", "null"] }
        amount: { type: number }
        date: { type: string }
        created_at: { type: string }
    ExpenseInput:
      type: object
      required: [category, amount, date]
      properties:
        category: { type: string }
        description: { type: string }
        amount: { type: number }
        date: { type: string }
    FinanceSummary:
      type: object
      required: [revenue, expenses, profit]
      properties:
        month: { type: ["string", "null"] }
        revenue: { type: number }
        expenses: { type: number }
        profit: { type: number }
        paid_invoices: { type: integer }
        unpaid_invoices: { type: integer }
        exempt_invoices: { type: integer }
        monthly_breakdown:
          type: array
          items:
            type: object
            required: [month, revenue, expenses]
            properties:
              month: { type: string }
              revenue: { type: number }
              expenses: { type: number }
...