# lib/api-zod/src/generated/types/expenseInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EDU SMART OS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ExpenseInput {
  category: string;
  description?: string;
  amount: number;
  date: string;
}